Echo cancellor interface changed to short (instead of float)
git-svn-id: http://svn.xiph.org/trunk/speex@6488 0101bb08-14d6-0310-b084-bc0e0c8e3800

+++ libspeex/speex_echo.h
@@ -73,7 +73,7 @@ SpeexEchoState *speex_echo_state_init(int frame_size, int filter_length);
 void speex_echo_state_destroy(SpeexEchoState *st);
 
 /** Performs echo cancellation a frame */
-void speex_echo_cancel(SpeexEchoState *st, float *ref, float *echo, float *out, float *Y);
+void speex_echo_cancel(SpeexEchoState *st, short *ref, short *echo, short *out, int *Y);
 
 #ifdef __cplusplus
 }
